refactor(api): move actions to modules/ as controller, model, service
